[username-name-case] username with non-alphanum or special characters
Postgres handling of usernames is the same as tables, clumn names, etc
(that is, tricky). Normally, they should consists of alphanum
characters and undrscores, and should not start with number and should
not coinside with keywords. In this case, in a stetement like

CREATE USER username;

username is transposed to lowercase before storing it, which creates
an illusion of them being case-insensitive. In reality, they are not.

In fact, when put inside double quotes, all limitations are
removed, but no transformation is done. So, it possible to have a user
with the name "SELECT", two different users with the names joe and
Joe, and the user "Hello, World!".

The logic to deal with these quirks was already implemented for table
names, columns names, etc. Here it is extended to usernames as well.