[issue-527] improvements to logics of capitalization of table columns in GDB
1. In the definition of the foreign table itself, always use the
original capitalization. This way, our implementation does not depend
on FDW implementation of the mapping between uppercase and lowercase
names between the postgres and the original database.
2. For each database type, we try to keep the original logic of the
columns naming. E.g., for Oracle, if the original column name was
written in a mix of capital and lowercase letters, it means that this
name was selected intentionally and we keep it. Otherwise, if it is
all uppercase, we assume the intended effective non-case-sensitive
behavior and reproduce it in GDB.