[jdbc-es-bugfix] Add selective columns query and update dependencies
Extended the query functionality to accept specific columns in the Requests.kt and DbService.kt files, instead of only querying all columns. Updated the MySQL JDBC Driver from version 8.1.0 to 8.2.0, and added the kotlinx-coroutines-core and kotlinx-coroutines-slf4j dependencies in the build.gradle file.

Also, increased precision for float type in jdbc.go (Elasticsearch return type as FLOAT with precision as 15 - which is DOUBLE, FLOAT has 6).