[misc-fixes] Update Elasticsearch JDBC driver and improve error checking
The codebase has been updated to use the latest version of the Elasticsearch JDBC driver (8.12.0). This involved changes in the build configuration files and the JDBC setup in various Golang files. Additionally, more rigorous file existence checks have been added to the script which handles DB2 library dependencies setup, improving the error messages that are displayed when certain dependency files are not found in the expected location.