Enhance Support for Monetary and Binary Data in Microsoft SQL Server
1. Introduce a Bytea<Vec<u8>> constructor within the Cell definition
to facilitate binary data handling.

2. Implement recognition for the monetary data type as transmitted by
the remote server. The Tiberius driver represents this data type as a
floating-point number. This value is then converted to the MONEOID
type in PostgreSQL, which is internally represented as an
integer. This conversion is achieved by multiplying the original value
by 100.

3. Binary data, initially represented as [u8], is converted to Vec<u8>
following the same conversion principle applied from &str to String.

4. Modify the original pagination query from "offset 0 rows fetch next
<offset + limit> rows only" to "TOP(<offset + limit>)". This
alteration is necessitated by the original query's requirement for an
'order by ...' condition, which may not always be specified by the
user. While the TOP statement does not permit specifying an offset, it
is noted that in the initial implementation, the offset is invariably
set to 0.

5. The incorporation of the Bytea constructor into the Cell
necessitates the addition of a corresponding case in the 'insert'
method within BigQuery. Given that this data type is presently
unsupported, it has been marked for future implementation with a
todo!() annotation.