escape column names as required by MS SQL server
Column names that contain spaces or other special symbols necessitate
escaping within Microsoft SQL Server environments. This is
accomplished by enclosing such column names within square
brackets. This practice is similar to the utilization of double quotes
for escaping in PostgreSQL; however, the use of square brackets for
this purpose is notably more prevalent in the context of Microsoft SQL
Server.